+++ lib/commands/initializer.dart
@@ -1,7 +1,8 @@
 import 'dart:io';
 
 import '../cached_build_runner.dart';
-import '../database/database_service.dart';
+import '../database/hive_database_service.dart';
+import '../database/redis_database_service.dart';
 import '../utils/utils.dart';
 
 /// This class configures the CachedBuildRunner as per the arguments parsed.
@@ -10,9 +11,7 @@ class Initializer {
   Future<CachedBuildRunner> init() async {
     /// the project directory is always where the `flutter run` command is executed
     /// which is the current directory
-    Utils.projectDirectory =
-        Platform.environment['CACHED_BUILD_RUNNER_PROJECT_DIRECTORY'] ??
-            Directory.current.path;
+    Utils.projectDirectory = Platform.environment['CACHED_BUILD_RUNNER_PROJECT_DIRECTORY'] ?? Directory.current.path;
 
     /// let's make the appCacheDirectory if not existing already
     Directory(Utils.appCacheDirectory).createSync(recursive: true);
@@ -21,9 +20,7 @@ class Initializer {
     Utils.initAppPackageName();
 
     /// initialize the database
-    final databaseService = Utils.isRedisUsed
-        ? RedisDatabaseService()
-        : HiveDatabaseService(Utils.appCacheDirectory);
+    final databaseService = Utils.isRedisUsed ? RedisDatabaseService() : HiveDatabaseService(Utils.appCacheDirectory);
     await databaseService.init();
 
     /// let's initiate the build