+++ lib/model/code_file.dart
@@ -2,15 +2,17 @@ class CodeFile {
   final String path;
   final String digest;
   final bool isTestFile;
+  final String? suffix;
 
   CodeFile({
     required this.path,
     required this.digest,
     this.isTestFile = false,
+    required this.suffix,
   });
 
   @override
   String toString() {
-    return '($path, $digest, $isTestFile)';
+    return '($path, $digest, $isTestFile, $suffix)';
   }
 }